Fix identical values being re-merged.
Look at the simple case of merging:

```javascript
{ }
{ foo: { bar: [ 3 ] } }
```

First the `foo` property will be looked at – because it doesn't exist in the parent object it will be set in the parent object. So now the parent object looks exactly the same as the child. However, the merge is not done; the other properties must be considered. So this results in `bar` being merged with the other copy of `bar` and so forth. Because this is not actually what we want (and because it introduces interesting side effects with arrays) there is now a short-circuit check to prevent identical objects from being merged.