Replace `inject` with `entry`.
The old `inject` helper was designed to muck around with the `entry` value in webpack configurations. This is fragile because it requires the consumer to tweak the `entry` value in the configuraiton itself and is not consistent with the other helpers. This changes now with the entry helper which mutates/replaces the `entry` value in consistent ways.

```javascript
// Set the `entry` to `index.js`
entry('index.js');

// Append `foo.js` to all existing entrypoints.
entry.append('foo.js');
entry((previous) => [...previous, 'foo.js'])
```

The `entry` function takes either a value to add to entry _or_ a function that maps the existing entry values to new ones. The values property is _always_ an array for consistency even though internally webpack can use objects, strings or arrays.

The callback has this signature:

```javascript
(previous: Array, key: ?String, config: Object) => { ... }
```

The `key` property represents the key in object-style entrypoints and `config` is the existing webpack configuration object.